+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Global toolbar.
+- Folder compare and sync between panes.
 
 ## [0.19.0] - 2026-06-20